version before shop section started. Cropper and image validation added in edit product also.

+++ glamyze/product_management/views.py
@@ -38,7 +38,7 @@ def product_details(request):
                 context['category_id'] = category_id
             elif subcategory_id:
                 context['subcategory_id'] = subcategory_id
-        return render(request,'admin/product.html',context)
+        return render(request,'my_admin/product.html',context)
     elif request.user.is_authenticated:
         return redirect('auth_app:home')
     else:
@@ -51,7 +51,6 @@ def product_add(request):
         context={'subcategories':subcategories}
         if request.POST:
             product_name = request.POST['product_name']
-            price = request.POST['price']
             subcategory_id = request.POST['subcategory']
             material = request.POST['material']
             color = request.POST['color']
@@ -70,15 +69,15 @@ def product_add(request):
             if error3:
                 errors.append(error3)
             if errors:
-                context.update({'product_name':product_name,'price':price,'subcategory_id':subcategory_id,'material':material,'color':color,'description':description,'errors':errors})
+                context.update({'product_name':product_name,'subcategory_id':subcategory_id,'material':material,'color':color,'description':description,'errors':errors})
                 return render(request,'admin/addproduct.html',context)
 
-            product_obj=Product(product_name=product_name,price=price,subcategory_id=subcategory_id,material=material,color=color,description=description,image1=image1,image2=image2,image3=image3)
+            product_obj=Product(product_name=product_name,subcategory_id=subcategory_id,material=material,color=color,description=description,image1=image1,image2=image2,image3=image3)
             product_obj.save()
             product_obj.refresh_from_db()
             request.session['product_id']=product_obj.id
             return redirect('product_management:product_size')
-        return render(request,'admin/addproduct.html',context)
+        return render(request,'my_admin/addproduct.html',context)
     elif request.user.is_authenticated:
         return redirect('auth_app:home')
     else:
@@ -90,6 +89,8 @@ def validate_image_format(image,image_name):
                 with Image.open(image) as img:
                     if img.format.lower() not in ['png', 'jpg', 'jpeg']:
                         return f"{image_name} must be a PNG or JPG file."
+                    else:
+                        return None
             except Exception:
                     return f"{image_name} is not a valid image file."
 
@@ -98,7 +99,7 @@ def validate_image_format(image,image_name):
 def product_size(request):
     if request.user.is_superuser:
         product_id = request.session.get('product_id')
-        obj = ProductSize.objects.filter(product_id=product_id)
+        obj = ProductVariant.objects.filter(product_id=product_id)
         excluded_sizes = obj.values_list('size_id', flat=True)
         available_sizes = Size.objects.exclude(id__in=excluded_sizes)
         if request.POST:
@@ -111,8 +112,12 @@ def product_size(request):
                         quantity = int(value) if value.strip() else 0
                     except ValueError:
                         quantity = 0 
-                    product_size_obj = ProductSize.objects.get(product_id=product_id,size_id=size_id)
+                    product_size_obj = ProductVariant.objects.get(product_id=product_id,size_id=size_id)
                     product_size_obj.quantity = quantity
+                    price_key = f'price{size_id}'
+                    price = request.POST.get(price_key, '0')
+                    price = float(price) if price.strip() else 0.0
+                    product_size_obj.price = price
                     product_size_obj.save()
                     updated = True
             if updated:
@@ -123,7 +128,7 @@ def product_size(request):
                 product_obj.save()
             del request.session['product_id']
             return redirect('product_management:product_details')
-        return render(request,'admin/productsize.html',{'product_varients':obj,'sizes':available_sizes})
+        return render(request,'my_admin/productsize.html',{'product_varients':obj,'sizes':available_sizes})
     elif request.user.is_authenticated:
         return redirect('auth_app:home')
     else:
@@ -135,7 +140,7 @@ def size_add(request):
         if request.POST:
             sizeid = request.POST['sizeid']
             product_id = request.session['product_id']
-            obj = ProductSize(size_id=sizeid,product_id=product_id)
+            obj = ProductVariant(size_id=sizeid,product_id=product_id)
             obj.save()
         return redirect('product_management:product_size')
     elif request.user.is_authenticated:
@@ -173,36 +178,57 @@ def product_edit(request,product_id):
     if request.user.is_superuser:    
         product = Product.objects.get(id=product_id)
         subcategories = SubCategory.objects.select_related('category')
-        if request.POST:
-            product.product_name = request.POST.get('product_name')
-            product.price = request.POST.get('price')
-            product.subcategory_id = request.POST.get('subcategory')
-            product.material = request.POST.get('material')
-            product.color = request.POST.get('color')
-            product.description = request.POST.get('description')
-            if request.FILES.get('image1'):
-                if product.image1:
-                    product.image1.delete()
-                product.image1 = request.FILES['image1']
-                
-            if request.FILES.get('image2'):
-                if product.image2:
-                    product.image2.delete()
-                product.image2 = request.FILES['image2']
-                
-            if request.FILES.get('image3'):
-                if product.image3:
-                    product.image3.delete()
-                product.image3 = request.FILES['image3']
-            product.save()
-            messages.success(request, 'Product updated successfully!')
-            return redirect('product_management:product_details')
         context = {
             'product': product,
             'subcategories' : subcategories
 
         }
-        return render(request, 'admin/editproducts.html', context)
+        if request.POST:
+            errors = []
+            if request.FILES.get('image1'):
+                image1 = request.FILES.get('image1')
+                error1 = validate_image_format(image1,'Image1')
+                if error1:
+                    errors.append(error1)
+                    
+                else:
+                    if product.image1:
+                        product.image1.delete()
+                    product.image1 = request.FILES['image1']
+                
+            if request.FILES.get('image2'):
+                image2 = request.FILES.get('image2')
+                error2 = validate_image_format(image2,'Image2')
+                if error2:
+                    errors.append(error2)
+                else:
+                    if product.image2:
+                        product.image2.delete()
+                    product.image2 = request.FILES['image2']
+                
+            if request.FILES.get('image3'):
+                image3 = request.FILES.get('image3')
+                error3 = validate_image_format(image3,'Image3')
+                if error3:
+                    errors.append(error3)
+                else:
+                    if product.image3:
+                        product.image3.delete()
+                    product.image3 = request.FILES['image3']
+            if errors:
+                print(errors)
+                context['errors'] = errors
+                return render(request, 'my_admin/editproducts.html',context)
+            product.product_name = request.POST.get('product_name')
+            product.subcategory_id = request.POST.get('subcategory')
+            product.material = request.POST.get('material')
+            product.color = request.POST.get('color')
+            product.description = request.POST.get('description')
+            product.save()
+            messages.success(request, 'Product updated successfully!')
+            return redirect('product_management:product_details')
+       
+        return render(request, 'my_admin/editproducts.html', context)
     elif request.user.is_authenticated:
             return redirect('auth_app:home')
     else:
@@ -210,31 +236,62 @@ def product_edit(request,product_id):
 
 
 @never_cache
-def product_new_size(request,product_id):
+def product_varient_management(request,product_id):
     if request.user.is_superuser:
-        obj = ProductSize.objects.filter(product_id=product_id)
+        
         if request.POST:
             for key, value in request.POST.items():
                 if key.isdigit() and value:
                     size_id = int(key) 
-                    obj = ProductSize(product_id=product_id,size_id=size_id)
-                    obj.save()
-            return redirect('product_management:product_details')
+                    if ProductVariant.objects.filter(product_id=product_id,size_id=size_id):
+                        pass
+                    else:
+                        obj = ProductVariant(product_id=product_id,size_id=size_id,is_listed=False)
+                        obj.save()
+        obj = ProductVariant.objects.filter(product_id=product_id)
         excluded_sizes = obj.values_list('size_id', flat=True)
-        added_size = Size.objects.filter(id__in=excluded_sizes)
+        product = Product.objects.get(id=product_id)
+        # Get added sizes with their listing status
+        added_sizes_with_status = []
+        for variant in obj:
+            added_sizes_with_status.append({
+                'size': variant.size,
+                'is_listed': variant.is_listed,
+                'price' : variant.price
+            })
+        
         available_sizes = Size.objects.exclude(id__in=excluded_sizes)
-        return render(request,'admin/newsize.html',{'added_sizes':added_size,'sizes':available_sizes})
+        return render(request,'my_admin/varient_mgmt.html',{'product':product, 'added_sizes':added_sizes_with_status,'sizes':available_sizes})
     elif request.user.is_authenticated:
         return redirect('auth_app:home')
     else:
         return redirect('auth_app:login')
+    
+def product_varient_management_post(request):
+    if request.user.is_superuser:
+        if request.POST:
+            product_id = request.POST['product_id']
+            variants = ProductVariant.objects.filter(product_id=product_id)
+            for variant in variants:
+                if request.POST.get(f'listed{variant.size.id}'):
+                    variant.is_listed = True
+                else:
+                    variant.is_listed = False
+                variant.price= request.POST.get(f'price{variant.size.id}')
+                variant.save()
+            return redirect('product_management:product_details')
+    elif request.user.is_authenticated:
+            return redirect('auth_app:home')
+    else:
+            return redirect('auth_app:login')
+
 
 
 @never_cache
 def product_add_quantity(request, product_id):
     if request.user.is_superuser:
         product = Product.objects.get(id=product_id)
-        product_sizes = ProductSize.objects.filter(product_id=product_id)
+        product_sizes = ProductVariant.objects.filter(product_id=product_id)
 
         if request.method == 'POST':
             updated_count = 0
@@ -254,7 +311,7 @@ def product_add_quantity(request, product_id):
                 messages.success(request, f'{updated_count} size(s) updated successfully!')
             return redirect('product_management:product_details')
 
-        return render(request, 'admin/addquantity.html', {'product_sizes': product_sizes, 'product': product})
+        return render(request, 'my_admin/addquantity.html', {'product_sizes': product_sizes, 'product': product})
     
     elif request.user.is_authenticated:
         return redirect('auth_app:home')
